requirements.txt
hdx-python-scraper==1.4.6

